chore(elements): Clean up form hooks [SDKI-587]
