+++ packages/elements/src/react/common/form/hooks/use-field-context.ts
@@ -1,0 +1,6 @@
+import * as React from 'react';
+
+import type { FieldDetails } from '~/internals/machines/form';
+
+export const FieldContext = React.createContext<Pick<FieldDetails, 'name'> | null>(null);
+export const useFieldContext = () => React.useContext(FieldContext);
