+++ packages/elements/src/react/common/form/hooks/use-field-feedback.ts
@@ -1,0 +1,9 @@
+import { type FieldDetails, fieldFeedbackSelector, useFormSelector } from '~/internals/machines/form';
+
+export function useFieldFeedback({ name }: Partial<Pick<FieldDetails, 'name'>>) {
+  const feedback = useFormSelector(fieldFeedbackSelector(name));
+
+  return {
+    feedback,
+  };
+}
