+++ packages/elements/src/react/common/form/hooks/use-global-errors.ts
@@ -1,0 +1,9 @@
+import { globalErrorsSelector, useFormSelector } from '~/internals/machines/form';
+
+export function useGlobalErrors() {
+  const errors = useFormSelector(globalErrorsSelector);
+
+  return {
+    errors,
+  };
+}
