+++ packages/elements/src/react/common/form/hooks/use-previous.ts
@@ -1,0 +1,11 @@
+import * as React from 'react';
+
+export function usePrevious<T>(value: T): T | undefined {
+  const ref = React.useRef<T>();
+
+  React.useEffect(() => {
+    ref.current = value;
+  }, [value]);
+
+  return ref.current;
+}
