+++ packages/elements/src/react/common/form/hooks/use-validity-state-context.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+
+export const ValidityStateContext = React.createContext<ValidityState | undefined>(undefined);
+export const useValidityStateContext = () => React.useContext(ValidityStateContext);
